requirements.txt
influxdb_client
python-dotenv
